fix warning 同步官方最新版部分代码
MEDIA_TYPE_URL was used from dlna_dmr_xiaodu, this is a deprecated constant which will be removed in HA Core 2025.10. Use MediaType.URL instead, please report it to the author of the 'dlna_dmr_xiaodu' custom integration

MEDIA_TYPE_VIDEO was used from dlna_dmr_xiaodu, this is a deprecated constant which will be removed in HA Core 2025.10. Use MediaType.VIDEO instead, please report it to the author of the 'dlna_dmr_xiaodu' custom integration

REPEAT_MODE_OFF was used from dlna_dmr_xiaodu, this is a deprecated constant which will be removed in HA Core 2025.10. Use RepeatMode.OFF instead, please report it to the author of the 'dlna_dmr_xiaodu' custom integration

REPEAT_MODE_ONE was used from dlna_dmr_xiaodu, this is a deprecated constant which will be removed in HA Core 2025.10. Use RepeatMode.ONE instead, please report it to the author of the 'dlna_dmr_xiaodu' custom integration

REPEAT_MODE_ALL was used from dlna_dmr_xiaodu, this is a deprecated constant which will be removed in HA Core 2025.10. Use RepeatMode.ALL instead, please report it to the author of the 'dlna_dmr_xiaodu' custom integration